+++ src/react-native/xcode.ts
@@ -7,6 +7,7 @@ import * as clack from '@clack/prompts';
 import chalk from 'chalk';
 import { makeCodeSnippet, showCopyPasteInstructions } from '../utils/clack';
 import { Project } from 'xcode';
+import * as Sentry from '@sentry/node';
 
 type BuildPhase = { shellScript: string };
 type BuildPhaseMap = Record<string, BuildPhase>;
@@ -66,65 +67,6 @@ export async function patchBundlePhase(
   );
 }
 
-export function unPatchBundlePhase(bundlePhase: BuildPhase | undefined) {
-  if (!bundlePhase) {
-    clack.log.warn(
-      `Could not find ${chalk.cyan(
-        'Bundle React Native code and images',
-      )} build phase.`,
-    );
-    return;
-  }
-
-  if (
-    !bundlePhase.shellScript.match(/sentry-cli\s+react-native\s+xcode/i) &&
-    !bundlePhase.shellScript.includes('sentry-xcode.sh')
-  ) {
-    clack.log.success(
-      `Build phase ${chalk.cyan(
-        'Bundle React Native code and images',
-      )} does not include Sentry.`,
-    );
-    return;
-  }
-
-  bundlePhase.shellScript = JSON.stringify(
-    removeSentryFromBundleShellScript(
-      <string>JSON.parse(bundlePhase.shellScript),
-    ),
-  );
-  clack.log.success(
-    `Build phase ${chalk.cyan(
-      'Bundle React Native code and images',
-    )} unpatched successfully.`,
-  );
-}
-
-export function removeSentryFromBundleShellScript(script: string): string {
-  return (
-    script
-      // remove sentry properties export
-      .replace(/^export SENTRY_PROPERTIES=sentry.properties\r?\n/m, '')
-      .replace(
-        /^\/bin\/sh .*?..\/node_modules\/@sentry\/react-native\/scripts\/collect-modules.sh"?\r?\n/m,
-        '',
-      )
-      // unwrap react-native-xcode.sh command.  In case someone replaced it
-      // entirely with the sentry-cli command we need to put the original
-      // version back in.
-      .replace(
-        /\.\.\/node_modules\/@sentry\/cli\/bin\/sentry-cli\s+react-native\s+xcode\s+\$REACT_NATIVE_XCODE/i,
-        '$REACT_NATIVE_XCODE',
-      )
-      .replace(
-        //  eslint-disable-next-line no-useless-escape
-        /\"\/bin\/sh.*?sentry-xcode.sh\s+\$REACT_NATIVE_XCODE/i,
-        // eslint-disable-next-line no-useless-escape
-        '"$REACT_NATIVE_XCODE',
-      )
-  );
-}
-
 export function findBundlePhase(buildPhases: BuildPhaseMap) {
   return Object.values(buildPhases).find((buildPhase) =>
     buildPhase.shellScript.match(/\/scripts\/react-native-xcode\.sh/i),
@@ -171,6 +113,9 @@ export function addSentryWithBundledScriptsToBundleShellScript(
         'Bundle React Native code and images',
       )} build phase.`,
     );
+    Sentry.captureException(
+      `Failed to patch 'Bundle React Native code and images' build phase.`,
+    );
     if (isLikelyExpoScript) {
       return new ErrorPatchSnippet(
         makeCodeSnippet(true, (unchanged, plus, _minus) => {
@@ -198,43 +143,6 @@ REACT_NATIVE_XCODE="$REACT_NATIVE_PATH/scripts/react-native-xcode.sh"
   }
 
   return patchedScript;
-}
-
-export function addSentryWithCliToBundleShellScript(
-  script: string,
-): string | ErrorPatchSnippet {
-  if (!script.includes('$REACT_NATIVE_XCODE')) {
-    clack.log.error(
-      `Could not find $REACT_NATIVE_XCODE in ${chalk.cyan(
-        'Bundle React Native code and images',
-      )} build phase. Skipping patching.`,
-    );
-    return new ErrorPatchSnippet(
-      makeCodeSnippet(true, (unchanged, plus, _minus) => {
-        return unchanged(`${plus(`export SENTRY_PROPERTIES=sentry.properties
-export EXTRA_PACKAGER_ARGS="--sourcemap-output $DERIVED_FILE_DIR/main.jsbundle.map"
-`)}
-/bin/sh -c "$WITH_ENVIRONMENT ${plus(
-          `\\"../node_modules/@sentry/cli/bin/sentry-cli react-native xcode`,
-        )} $REACT_NATIVE_XCODE${plus(`\\"`)}"
-${plus(
-  `/bin/sh -c "$WITH_ENVIRONMENT ../node_modules/@sentry/react-native/scripts/collect-modules.sh"`,
-)}
-`);
-      }),
-    );
-  }
-  return (
-    'export SENTRY_PROPERTIES=sentry.properties\n' +
-    'export EXTRA_PACKAGER_ARGS="--sourcemap-output $DERIVED_FILE_DIR/main.jsbundle.map"\n' +
-    script.replace(
-      '$REACT_NATIVE_XCODE',
-      () =>
-        // eslint-disable-next-line no-useless-escape
-        '\\"../node_modules/@sentry/cli/bin/sentry-cli react-native xcode $REACT_NATIVE_XCODE\\"',
-    ) +
-    '\n/bin/sh -c "$WITH_ENVIRONMENT ../node_modules/@sentry/react-native/scripts/collect-modules.sh"\n'
-  );
 }
 
 export function addDebugFilesUploadPhaseWithBundledScripts(
@@ -266,83 +174,6 @@ export function addDebugFilesUploadPhaseWithBundledScripts(
   );
 }
 
-export function addDebugFilesUploadPhaseWithCli(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  xcodeProject: any,
-  { debugFilesUploadPhaseExists }: { debugFilesUploadPhaseExists: boolean },
-) {
-  if (debugFilesUploadPhaseExists) {
-    clack.log.warn(
-      `Build phase ${chalk.cyan(
-        'Upload Debug Symbols to Sentry',
-      )} already exists.`,
-    );
-    return;
-  }
-
-  xcodeProject.addBuildPhase(
-    [],
-    'PBXShellScriptBuildPhase',
-    'Upload Debug Symbols to Sentry',
-    null,
-    {
-      shellPath: '/bin/sh',
-      shellScript: `
-WITH_ENVIRONMENT="../node_modules/react-native/scripts/xcode/with-environment.sh"
-if [ -f "$WITH_ENVIRONMENT" ]; then
-  . "$WITH_ENVIRONMENT"
-fi
-export SENTRY_PROPERTIES=sentry.properties
-[ "$SENTRY_INCLUDE_NATIVE_SOURCES" = "true" ] && INCLUDE_SOURCES_FLAG="--include-sources" || INCLUDE_SOURCES_FLAG=""
-../node_modules/@sentry/cli/bin/sentry-cli debug-files upload "$INCLUDE_SOURCES_FLAG" "$DWARF_DSYM_FOLDER_PATH"
-`,
-    },
-  );
-  clack.log.success(
-    `Added Build phase ${chalk.cyan('Upload Debug Symbols to Sentry')}.`,
-  );
-}
-
-export function unPatchDebugFilesUploadPhase(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  xcodeProject: any,
-) {
-  const buildPhasesMap =
-    xcodeProject.hash.project.objects.PBXShellScriptBuildPhase || {};
-
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-  const debugFilesUploadPhaseResult = findDebugFilesUploadPhase(buildPhasesMap);
-  if (!debugFilesUploadPhaseResult) {
-    clack.log.success(
-      `Build phase ${chalk.cyan('Upload Debug Symbols to Sentry')} not found.`,
-    );
-    return;
-  }
-
-  const [debugFilesUploadPhaseKey] = debugFilesUploadPhaseResult;
-  const firstTarget: string = xcodeProject.getFirstTarget().uuid;
-  const nativeTargets = xcodeProject.hash.project.objects.PBXNativeTarget;
-
-  // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-  delete buildPhasesMap[debugFilesUploadPhaseKey];
-  // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-  delete buildPhasesMap[`${debugFilesUploadPhaseKey}_comment`];
-  const phases = nativeTargets[firstTarget].buildPhases;
-  if (phases) {
-    for (let i = 0; i < phases.length; i++) {
-      if (phases[i].value === debugFilesUploadPhaseKey) {
-        phases.splice(i, 1);
-        break;
-      }
-    }
-  }
-  clack.log.success(
-    `Build phase ${chalk.cyan(
-      'Upload Debug Symbols to Sentry',
-    )} removed successfully.`,
-  );
-}
-
 export function findDebugFilesUploadPhase(
   buildPhasesMap: Record<string, BuildPhase>,
 ): [key: string, buildPhase: BuildPhase] | undefined {
@@ -363,14 +194,23 @@ export function writeXcodeProject(
   xcodeProjectPath: string,
   xcodeProject: Project,
 ) {
-  const newContent = xcodeProject.writeSync();
-  const currentContent = fs.readFileSync(xcodeProjectPath, 'utf-8');
-  if (newContent === currentContent) {
-    return;
-  }
+  try {
+    const newContent = xcodeProject.writeSync();
+    const currentContent = fs.readFileSync(xcodeProjectPath, 'utf-8');
+    if (newContent === currentContent) {
+      return;
+    }
 
-  fs.writeFileSync(xcodeProjectPath, newContent, 'utf-8');
-  clack.log.success(
-    chalk.green(`Xcode project ${chalk.cyan(xcodeProjectPath)} changes saved.`),
-  );
+    fs.writeFileSync(xcodeProjectPath, newContent, 'utf-8');
+    clack.log.success(
+      chalk.green(
+        `Xcode project ${chalk.cyan(xcodeProjectPath)} changes saved.`,
+      ),
+    );
+  } catch (error) {
+    clack.log.error(
+      `Error while writing Xcode project ${chalk.cyan(xcodeProjectPath)}`,
+    );
+    Sentry.captureException('Error while writing Xcode project');
+  }
 }